Allow files to load inspite of decode errors. - remove unnecesary code inside the NEED_EMSTRING_WCHAR_T defined - add utf-16 code, buts its not hooked in yet

+++ Editor/Include/Unix/os_unix.h
@@ -60,4 +60,4 @@ const int DBG_NO_DBG_MSG    (0x00001000&DBG___ON);  // always defined
 
 extern void OutputDebugString( const char *message );
 
-typedef unsigned short EmacsChar_t;
+typedef unsigned int EmacsChar_t;
